+++ mrpt_bridge/include/mrpt_bridge/stereo_image.h
@@ -1,6 +1,17 @@
-//
-// Created by raghavender on 21/08/17.
-//
+/* +------------------------------------------------------------------------+
+   |                     Mobile Robot Programming Toolkit (MRPT)            |
+   |                          http://www.mrpt.org/                          |
+   |                                                                        |
+   | Copyright (c) 2005-2017, Individual contributors, see AUTHORS file     |
+   | See: http://www.mrpt.org/Authors - All rights reserved.                |
+   | Released under BSD License. See details in http://www.mrpt.org/License |
+   +------------------------------------------------------------------------+ */
+
+/*---------------------------------------------------------------
+	APPLICATION: mrpt_ros bridge
+	FILE: stereo_image.h
+	AUTHOR: Raghavender Sahdev <raghavendersahdev@gmail.com>
+  ---------------------------------------------------------------*/
 
 #ifndef MRPT_BRIDGE_STEREO_IMAGE_H
 #define MRPT_BRIDGE_STEREO_IMAGE_H